tests/wpt/web-platform-tests/css/css-masking/parsing/mask-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-masking-1/#the-mask">
<meta name="assert" content="mask supports only the '<mask-layer>#' grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('mask', 'none linear-gradient(to 0% 0%, rgb(0, 0, 0), rgb(0, 0, 255))');
test_invalid_value('mask', 'none alpha luminance');
test_invalid_value('mask', 'linear-gradient(to 0% 0%, rgb(0, 0, 0), rgb(0, 0, 255)) url("https://example.com/")');

test_invalid_value('mask', '1px 2px 3px');
test_invalid_value('mask', '1px 2px 3px 4px');
test_invalid_value('mask', '1px 2px / contain / cover');

test_invalid_value('mask', 'repeat-y repeat-x');

test_invalid_value('mask', 'stroke-box no-clip view-box');

test_invalid_value('mask', 'border-box view-box padding-box');

test_invalid_value('mask', 'no-clip border-box content-box');

test_invalid_value('mask', 'content-box stroke-box no-clip');

test_invalid_value('mask', 'add intersect');
</script>
</body>
</html>
